refactor: consolidate locale-cookie constants into @plumix/core/i18n
Extract the duplicated `ADMIN_LOCALE_COOKIE`, `ADMIN_LOCALE_COOKIE_PATH`, `ONE_YEAR_SECONDS` and
`buildLocaleCookie` builder into `packages/core/src/i18n/cookie.ts`, re-exported via the existing
`@plumix/core/i18n` subpath. Before, the constants and builder lived in three places — the SSR
resolver in `runtime/admin-shell.ts`, the `user.setLocale` RPC, and the pre-auth login dropdown
in admin — guarded by a lockstep invariant comment. The three writers/readers now converge on
one module so drift between pre-auth and post-auth persistence is structurally impossible.

A new colocated `cookie.test.ts` pins the byte-string contract (with + without `Secure`); admin's
`-locale-param.test.ts` keeps `nextSearchForLang` + `writeLocaleCookie` and imports
`buildLocaleCookie` from `@plumix/core/i18n` for the round-trip assertion. No behavior change.

Fixes #803
